requirements.txt
amqp==1.4.9
ansible==2.3.2.0
anyjson==0.3.3
asn1crypto==0.22.0
bcrypt==3.1.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cipher==0.1
configparser==3.5.0
confusable-homoglyphs==2.0.2
crypto==1.4.1
cryptography==2.0.3
diff-match-patch==20121119
Django==1.11.6
django-bootstrap3==9.0.0
django-breadcrumbs==1.1.4a1
django-celery==3.2.1
django-crispy-forms==1.6.1
django-crispy-forms-registration==0.1.3
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.5.1
django-paginator==0.2a3
django-registration==2.2
django-reversion==2.0.10
django-xadmin==0.5.0
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
httplib2==0.9.2
idna==2.6
ipaddress==1.0.18
IPy==0.83
jdcal==1.3
jenkins==1.0.2
Jinja2==2.9.6
kombu==3.0.37
ldap==1.0.2
ldap3==2.3
MarkupSafe==1.0
Naked==0.1.31
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
paramiko==2.2.1
Pillow==4.3.0
pkg-resources==0.0.0
pyasn1==0.3.3
pycparser==2.18
pycrypto==2.6.1
PyMySQL==0.7.11
PyNaCl==1.1.2
pytz==2017.2
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
six==1.10.0
tablib==0.12.1
toml==0.9.3
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0

